% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumer.R
\name{Consumer}
\alias{Consumer}
\title{Consumer object}
\description{
The Consumer object consumes messages sent to the message
broker and spawns Workers to run functions in background.
}
\section{Usage}{

\preformatted{
dummy1 <- function(){Sys.sleep(5); print('dummy1')}
dummy2 <- function(){Sys.sleep(5); print('dummy2')}

tasks <- Tasks$new(tasklist=list(dummy1=dummy1, dummy2=dummy2))
queue <- Queue$new()

consumer <- Consumer$new(queue=queue, nworkers=2, tasks=tasks)
consumer$consume()
}
}

\section{Arguments}{

\describe{
    \item{consumer:}{A \code{Consumer} object.}
    \item{tasks:}{A \code{Tasks} object.}
    \item{queue:}{A \code{Queue} object.}
    \item{backend_url:}{A url string of the following format: "provider://host:port".}
    \item{nworkers:}{Numeric. Maximum number of background workers.
                    Default is 1.}
}
}

