% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{getSNNiGraph}
\alias{getSNNiGraph}
\title{Get and Convert SNN Graph object into igraph object}
\usage{
getSNNiGraph(data)
}
\arguments{
\item{data}{\code{Seurat} object}
}
\value{
boolean indicating if SNN Graph object exists
}
\description{
This is used to facilitate leiden clustering.
}
