% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\title{Queue object}
\description{
This object establishes an interface for different
message brokers using the AMQP Protocol.
}
\section{Usage}{

\preformatted{queue <- Queue$new(provider='redis', host='localhost',
                                   port=6379, name='celery')
msg <- queue$pull()
queue$push(msg)
}
}

\section{Arguments}{

\describe{
    \item{queue:}{A \code{Queue} object.}
    \item{provider:}{Character scalar referring to the message broker provider.
                    \code{'redis'} is the only available option until now.}
    \item{host:}{Character. Message broker instance address.}
    \item{port:}{Numeric. Message broker port.}
}
}

