% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_backend.R
\name{RedisBackend}
\alias{RedisBackend}
\title{RedisBackend object}
\arguments{
\item{host}{Character. Message broker instance address.}

\item{port}{Numeric. Message broker port.}

\item{username}{Ignored (included for consistency with other non-Redis queues)}

\item{password}{Character. Redis password.}

\item{db}{Numeric. Database.}
}
\description{
This object establishes an interface with Redis as a results backend.
}
\section{Usage}{
\preformatted{backend <- RedisBackend$new(host='localhost', port=6379, password=NULL, db=0)
backend$store_result(123, TRUE)
}
}

