% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis_queue.R
\name{RedisQueue}
\alias{RedisQueue}
\title{RedisQueue object}
\arguments{
\item{name}{The name of the queue.}

\item{host}{Message broker instance address.}

\item{port}{Message broker port.}

\item{password}{Redis password}

\item{db}{Database number}
}
\description{
This object establishes an interface for Redis as a message broker
}
\section{Usage}{
\preformatted{queue <- RedisQueue$new(host='localhost',
                        port=6379, name='celery',
                        password=NULL, db=0)
msg <- queue$pull()
queue$push(msg)
}
}

