% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rworker.R
\name{Rworker}
\alias{Rworker}
\title{Rworker object}
\description{
The Rworker object consumes messages sent to the message
broker and sends tasks to be executed in background by the
process pool.
}
\section{Usage}{

\preformatted{
 \dontrun{ 
       # Create rworker instance
       rwork <- rworker()

       # Register task
       (function() { 
           Sys.sleep(5)
       }) %>% rwork$task(name='long_running_task')

       # Send task for background execution
       rwork$execute('long_running_task')

       # Listen to messages from message queue
       rwork$consume()
 }

}
}

\section{Arguments}{

\describe{
    \item{qname:}{The name of the message queue.}
    \item{workers:}{The number of background worker processes.}
    \item{queue:}{A url string of type "provider://host:port".}
    \item{backend:}{A url string of type "provider://host:port".}
}
}

