% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rworker.R
\name{Tasks}
\alias{Tasks}
\title{Tasks object}
\description{
This object is stores all the functions that can be called by the
message broker.
}
\section{Usage}{

\preformatted{
dummy1 <- function(){Sys.sleep(5); print('dummy1')}
dummy2 <- function(){Sys.sleep(5); print('dummy2')}

tasks <- Tasks$new(tasklist=list(dummy1, dummy2)

}
}

\section{Arguments}{

\describe{
    \item{tasks:}{A \code{Tasks} object.}
    \item{tasklist:}{List object containing all functions.}
    \item{src:}{Character. Name of file containing functions.}
}
}

