% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ter.R
\name{TER}
\alias{TER}
\title{Task Execution Request}
\description{
Task Execution Request
}
\section{Usage}{
\preformatted{task_request <- ter(message)
}
}

\section{Details}{

The TER object represents the request sent by Celery.
}

