% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{Worker}
\alias{Worker}
\alias{Worker}
\alias{worker}
\title{Worker object}
\usage{
worker()
}
\description{
This object is responsible for running tasks consumed from the
message broker. It uses processx to spawn background processes.
}
\section{Usage}{

\dontrun{
\preformatted{
 w <- worker()

 w$listen()
}
}
}

