% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rworker.R
\name{Worker}
\alias{Worker}
\title{Worker object}
\description{
This object is responsible for running tasks consumed from the
message broker. It uses processx to spawn background processes.
}
\section{Usage}{

\preformatted{
dummy1 <- function(){Sys.sleep(5); print('dummy1')}
dummy2 <- function(){Sys.sleep(5); print('dummy2')}

tasks <- Tasks$new(tasklist=list(dummy1, dummy2)

worker1 <- Worker$new(FUN='dummy1',tasks=tasks)
worker2 <- Worker$new(FUN='dummy2',tasks=tasks)
}
}

\section{Arguments}{

\describe{
    \item{worker:}{A \code{Worker} object.}
    \item{FUN:}{Character. Name of the function to be executed in background.}
    \item{params:}{List object mapping parameters to values.
                  E.g. list('param'='value').}
    \item{tasks:}{A \code{Tasks} object.}
}
}

