% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{Worker}
\alias{Worker}
\alias{worker}
\title{Worker object}
\usage{
worker()
}
\description{
This object listens for tasks sent by some Rworker instance.
After running, the status of the execution is sent back
to the Rworker instance.
}
\section{Usage}{
\preformatted{w <- worker()
}
}

\section{Details}{

\code{$listen()} listens for tasks sent by the Rworker instance
\code{$report()} returns the task execution state to Rworker instance
}

\examples{
\dontrun{
w <- worker()
w$listen()
}
}
